Break up the file source
This commit breaks up the file source in a fashion which drops
FPs that are bound to paths which no longer exist, rotate faster
and with simpler code. Unfortunately we've changed how the file_watcher
read_line works and this has made the QC test for file_watcher sit
and spin. The next commit will have to correct that.
